refactor: moved icon definitions to constants
