+++ src/constants/icons.ts
@@ -1,0 +1,3 @@
+export const BLOOD_DECK: string = "🟥";
+export const ELIMINATED: string = "💀";
+export const SAND_DECK: string = "🟨";
